#20 add-card 폴더 구조 변경
