+++ components/add-card/bollingerband/bollingerband-singal-row.tsx
