+++ components/add-card/sma/sma-condition-content.tsx
